refactor(rent): move use cases into rent folder
