+++ src/main/java/com/medspace/application/usecase/rent/review/DeleteReviewByIdUseCase.java
@@ -1,4 +1,4 @@
-package com.medspace.application.usecase.review;
+package com.medspace.application.usecase.rent.review;
 
 import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Review;
